fix: Change appendBook logic
AppendBook will now append on load, and when a book is saved. This
differs to before where all books were appended each time the saved
books view was loaded. This means we had to reset the saved view section
each time we load it. Now because appendBook only runs on load and when
a book is saved, we do not need to reset.